+++ recipe/deploy/prepare.php
@@ -7,6 +7,7 @@
 
 namespace Deployer;
 
+use Deployer\Exception\GracefulShutdownException;
 use function Deployer\Support\str_contains;
 
 desc('Preparing host for deploy');
@@ -36,4 +37,20 @@ task('deploy:prepare', function () {
 
     // Create shared dir.
     run("cd {{deploy_path}} && if [ ! -d shared ]; then mkdir shared; fi");
+
+    // Check and save the remote HEAD/revision and compare it with the existing saved one (if any)
+    // This avoid unnecessary releases when the last commit id matches the existing one (HEAD)
+    $repository  = trim(get('repository'));
+    $revision    = input()->getOption('revision') ?? null;
+    $remoteHead  = $revision ?? run(sprintf('%s ls-remote %s HEAD | tr -d "HEAD"', get('bin/git'), $repository));
+
+    if (true === get('check_remote_head') && null == input()->getOption('tag')) {
+        $headPath = trim(get('deploy_path').'/.dep/HEAD');
+        $headContents = run(sprintf('test -e %s && cat %1$s', $headPath));
+        //check if HEAD file is exists and then compare it
+        if (trim($headContents) === trim($remoteHead)) {
+            throw new GracefulShutdownException("Already up-to-date.");
+        }
+    }
+    run("cd {{deploy_path}} && echo ".$remoteHead.' > .dep/HEAD');
 });