+++ CHANGELOG.md
@@ -68,6 +68,7 @@
 - Support to define remote shell path via host-config [#1708] [#1709] [#1709]
 - Added `horizon:terminate` to the Laravel recipe
 - Added `migrations_config` option to the Symfony recipes to specify Doctrine migration configuration to use
+- Added `check_remote_head` option, by setting this to true, deployer will avoid unnecessary new releases by checking the remote git HEAD without cloning the repo [#1755]
 - Added recipe for sulu 2.0 [#1758]
 - Added recipe for sulu 1.x and improve sulu 2.0 recipe [#1764]
 - Added `become` option for rsync upload
@@ -501,6 +502,7 @@
 [#1775]: https://github.com/deployphp/deployer/pull/1775
 [#1764]: https://github.com/deployphp/deployer/pull/1764
 [#1758]: https://github.com/deployphp/deployer/pull/1758
+[#1755]: https://github.com/deployphp/deployer/pull/1755
 [#1709]: https://github.com/deployphp/deployer/issues/1709
 [#1708]: https://github.com/deployphp/deployer/pull/1708
 [#1677]: https://github.com/deployphp/deployer/pull/1677